requirements.txt
fastapi
sqlalchemy
uvicorn
psycopg2-binary
pydantic
python-dotenv
alembic
